+++ src/main/java/me/study/shop/repository/ProductRepository.java
@@ -1,5 +1,6 @@
-package me.study.shop.domain;
+package me.study.shop.repository;
 
+import me.study.shop.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
